refactor: modularise dbapi into db package
