+++ tahrir_api/db/__init__.py
@@ -1,0 +1,13 @@
+# ruff: noqa: F401
+from .assertion import AssertionMethod
+from .authorization import AuthorizationMethod
+from .badge import BadgeMethod
+from .invitation import InvitationMethod
+from .issuer import IssuerMethod
+from .leaderboard import LeaderboardMethod
+from .milestone import MilestoneMethod
+from .person import PersonMethod
+from .rarity import RarityMethod
+from .series import SeriesMethod
+from .tag import TagMethod
+from .team import TeamMethod
